requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Django==5.2.7
django-autoslug==1.9.9
django-cors-headers==4.9.0
django-dirtyfields==1.9.7
django-phonenumber-field==8.3.0
django-phonenumbers==1.0.1
django-sendgrid-v5==1.3.0
django-versatileimagefield==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
frozenlist==1.8.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
kombu==5.5.4
MarkupSafe==3.0.3
multidict==6.7.0
oauthlib==3.3.1
openai==2.7.2
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
redis==5.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
twilio==9.8.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
yarl==1.22.0
PyMuPDF==1.26.6
pdfplumber==0.11.8
pypdf==6.4.0
pdf2image==1.17.0
pytesseract==0.3.13
weasyprint==66.0
Jinja2==3.1.6
jsonschema==4.25.1
PyPDF2==3.0.1
django-filter==25.2
pandas==3.0.0
openpyxl==3.1.5
